% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_.R
\name{change_SCOPE_options}
\alias{change_SCOPE_options}
\title{a function change the current SCOPE options}
\usage{
change_SCOPE_options(
  patch = "D:/model/STEMMUS_SCOPE/",
  site_name = NA,
  run_name = NA,
  calc_ebal = 1,
  calc_vert_profiles = 1,
  calc_fluor = 1,
  calc_planck = 0,
  calc_directional = 0,
  calc_xanthophyllabs = 1,
  calc_PSI = 0,
  rt_thermal = 0,
  calc_zo = 1,
  soilspectrum = 0,
  soil_heat_method = 1,
  Fluorescence_model = 0,
  calc_rss_rbs = 0,
  apply_T_corr = 1,
  verify = 0,
  save_headers = 1,
  makeplots = 0,
  simulation = 1
)
}
\arguments{
\item{patch}{the patch to the STEMMUS_SCOPE model directory, default "D:/model/STEMMUS_SCOPE/",}

\item{site_name, run_name}{the name of the location and the name of run. The last can be used to name runs with different model parameters or settings,}

\item{calc_ebal, calc_vert_profiles, calc_fluor, calc_planck, calc_directional, calc_xanthophyllabs, calc_PSI, rt_thermal, calc_zo, soilspectrum, soil_heat_method, Fluorescence_model, calc_rss_rbs, apply_T_corr, verify, save_headers, makeplots, simulation}{default c(1,1,1,0,0,1,0,0,1,0,1,0,0,1,0,1,0,1)}
}
\value{
write the new SCOPE model options in the input_data.xlsx file
}
\description{
\code{change_SCOPE_options} change current SCOPE model options in the input_data.xlsx file
}
\details{
This function change current SCOPE model options selected to run, @seealso \code{\link[=info_SCOPE_options]{info_SCOPE_options()}}
}
\examples{

change_SCOPE_options(site_name = "DE-Hai",
                     run_name = "ECdata_36",
                     calc_ebal = 1)

formattable::formattable(check_SCOPE_options(site_name = "DE-Hai", run_name = "ECdata_36"),
                         align = c("c","l"))

}
\seealso{
Other check the current parameters and settings: 
\code{\link{change_SCOPE_constants}()},
\code{\link{change_STEMMUS_DefineConstants}()},
\code{\link{change_STEMMUS_ModelSettings}()},
\code{\link{change_STEMMUS_SoilConstants}()},
\code{\link{check_SCOPE_constants}()},
\code{\link{check_SCOPE_options}()},
\code{\link{check_SCOPE_tsInputs}()},
\code{\link{check_STEMMUS_DefineConstants}()},
\code{\link{check_STEMMUS_ForcingGlobal}()},
\code{\link{check_STEMMUS_ModelSettings}()},
\code{\link{check_STEMMUS_SoilConstants}()},
\code{\link{check_STEMMUS_SoilInitials}()},
\code{\link{check_STEMMUS_SoilProperties}()}
}
\concept{check the current parameters and settings}
