% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_.R
\name{change_STEMMUS_DefineConstants}
\alias{change_STEMMUS_DefineConstants}
\title{a function to change the current STEMMUS Model define Constants file}
\usage{
change_STEMMUS_DefineConstants(
  patch = "D:/model/STEMMUS_SCOPE/",
  A = "6.02214E23",
  h = "6.6262E-34",
  c = "299792458",
  cp = "1004",
  cp_specific = "1.013E-3",
  R = "8.314",
  R_specific = "0.287",
  rhoa = "1.2047",
  kappa = "0.4",
  MH2O = "18",
  Mair = "28.96",
  MCO2 = "44",
  sigmaSB = "5.67E-8",
  deg2rad = "pi / 180",
  C2K = "273.15",
  CKTN = "(50 + 2.575 * 20)",
  l = "0.5",
  g = "981",
  RHOL = "1",
  RHOI = "0.92",
  Rv = "461.5 * 1e4",
  RDA = "287.1 * 1e4",
  RHO_bulk = "1.25",
  Hc = "0.02",
  GWT = "7",
  MU_a = "1.846 * 10^(-4)",
  Gamma0 = "71.89",
  Gamma_w = "const.RHOL*const.g",
  Lambda1 = "0.228 / 100",
  Lambda2 = "-2.406/100",
  Lambda3 = "4.909/100",
  MU_W0 = "2.4152 * 10^(-4)",
  MU1 = "4742.8",
  b = "4 * 10^(-6)",
  W0 = "1.001 * 10^3",
  c_L = "4.186",
  c_V = "1.870",
  c_a = "1.005",
  c_i = "2.0455",
  lambdav = "2.45",
  k = "0.41"
)
}
\arguments{
\item{patch}{the patch to the STEMMUS_SCOPE model directory, default "D:/model/STEMMUS_SCOPE/",}
}
\value{
write the new STEMMUS define Constants inputs
}
\description{
\code{change_STEMMUS_DefineConstants} change current model define Constants inputs for STEMMUS
}
\details{
This function change current STEMMUS model define Constants inputs @seealso \code{\link[=check_STEMMUS_DefineConstants]{check_STEMMUS_DefineConstants()}}.
\strong{important} differently from SCOPE constant parameters any change in the constants will impact all
next runs and not a specific site_name/run_name run as in SCOPE.
}
\examples{


change_define_constants(A = "6.02214E23")

formattable::formattable(check_STEMMUS_DefineConstants(), align = c("l","c"))

}
\seealso{
Other check the current parameters and settings: 
\code{\link{change_SCOPE_constants}()},
\code{\link{change_SCOPE_options}()},
\code{\link{change_STEMMUS_ModelSettings}()},
\code{\link{change_STEMMUS_SoilConstants}()},
\code{\link{check_SCOPE_constants}()},
\code{\link{check_SCOPE_options}()},
\code{\link{check_SCOPE_tsInputs}()},
\code{\link{check_STEMMUS_DefineConstants}()},
\code{\link{check_STEMMUS_ForcingGlobal}()},
\code{\link{check_STEMMUS_ModelSettings}()},
\code{\link{check_STEMMUS_SoilConstants}()},
\code{\link{check_STEMMUS_SoilInitials}()},
\code{\link{check_STEMMUS_SoilProperties}()}
}
\concept{check the current parameters and settings}
