% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_.R
\name{change_STEMMUS_SoilConstants}
\alias{change_STEMMUS_SoilConstants}
\title{to change the current STEMMUS Model Soil Constants}
\usage{
change_STEMMUS_SoilConstants(
  patch = "D:/model/STEMMUS_SCOPE/",
  hd = -1e+07,
  hm = -9899,
  CHST = 0,
  Elmn_Lnth = 0,
  Dmark = 0,
  Phi_S1 = -17.9,
  Phi_S2 = -17,
  Phi_S3 = -17,
  Phi_S4 = -19,
  Phi_S5 = -10,
  Phi_S6 = -10,
  Phi_soc = -0.0103,
  Lamda_soc = 2.7,
  Theta_soc = 0.6,
  XK = 0.025
)
}
\arguments{
\item{patch}{the patch to the STEMMUS_SCOPE model directory, default "D:/model/STEMMUS_SCOPE/",}

\item{hd, hm, CHST, Elmn_Lnth, Dmark, Phi_S1, Phi_S2, Phi_S3, Phi_S4, Phi_S5, Phi_S6, Phi_soc, Lamda_soc, Theta_soc, XK, }{dafault = c(-1e7, -9899,0,0,0,-17.9,-17,-17,-19,-10,-10,-0.0103,2.7,0.6,0.025)}
}
\value{
write the new STEMMUS Soil Constants inputs
}
\description{
\code{change_STEMMUS_SoilConstants} change current model Soil Constants inputs for STEMMUS
}
\details{
This function change current STEMMUS model Soil Constants inputs @seealso \code{\link[=info_STEMMUS_SoilConstants]{info_STEMMUS_SoilConstants()}}.
\strong{important} differently from SCOPE constant parameters any change in the constants will impact all
next runs and not a specific site_name/run_name run as in SCOPE.
}
\examples{

\dontrun{
change_soil_constants(XK = 0.025)

formattable::formattable(check_STEMMUS_SoilConstants(), align = c("l","c"))
}

}
\seealso{
Other check the current parameters and settings: 
\code{\link{change_SCOPE_constants}()},
\code{\link{change_SCOPE_options}()},
\code{\link{change_STEMMUS_DefineConstants}()},
\code{\link{change_STEMMUS_ModelSettings}()},
\code{\link{check_SCOPE_constants}()},
\code{\link{check_SCOPE_options}()},
\code{\link{check_SCOPE_tsInputs}()},
\code{\link{check_STEMMUS_DefineConstants}()},
\code{\link{check_STEMMUS_ForcingGlobal}()},
\code{\link{check_STEMMUS_ModelSettings}()},
\code{\link{check_STEMMUS_SoilConstants}()},
\code{\link{check_STEMMUS_SoilInitials}()},
\code{\link{check_STEMMUS_SoilProperties}()}
}
\concept{check the current parameters and settings}
