% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_.R
\name{check_SCOPE_constants}
\alias{check_SCOPE_constants}
\title{to check current SCOPE constants}
\usage{
check_SCOPE_constants(
  patch = "D:/model/rSTEMMUS_SCOPE/",
  site_name = NA,
  run_name = NA
)
}
\arguments{
\item{patch}{the patch to the STEMMUS_SCOPE model directory, default "D:/model/rSTEMMUS_SCOPE/",}

\item{site_name, run_name}{the name of the location and the name of run. The last can be used to name runs with different model parameters or settings,}
}
\value{
a table with information about SCOPE current model constant parameters
}
\description{
\code{check_SCOPE_constants} present current SCOPE model constant filled in the input_data.xlsx file
}
\details{
This function shows current SCOPE  constant used to run the model. If it was not changed while
using @seealso \code{\link[=input_constants]{input_constants()}} it will present the default values.
}
\examples{

\dontrun{
formattable::formattable(check_SCOPE_constants(site_name = "DE-Hai", run_name = "ECdata_36")[[1]],
            align = c("l","c","c","l"))

formattable::formattable(check_SCOPE_constants(site_name = "DE-Hai", run_name = "ECdata_36")[[2]],
            align = c("l","c","c","l"))
}

}
\seealso{
Other check the current parameters and settings: 
\code{\link{change_SCOPE_constants}()},
\code{\link{change_SCOPE_options}()},
\code{\link{change_STEMMUS_DefineConstants}()},
\code{\link{change_STEMMUS_ModelSettings}()},
\code{\link{change_STEMMUS_SoilConstants}()},
\code{\link{check_SCOPE_options}()},
\code{\link{check_SCOPE_tsInputs}()},
\code{\link{check_STEMMUS_DefineConstants}()},
\code{\link{check_STEMMUS_ForcingGlobal}()},
\code{\link{check_STEMMUS_ModelSettings}()},
\code{\link{check_STEMMUS_SoilConstants}()},
\code{\link{check_STEMMUS_SoilInitials}()},
\code{\link{check_STEMMUS_SoilProperties}()}
}
\concept{check the current parameters and settings}
