% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_.R
\name{check_SCOPE_options}
\alias{check_SCOPE_options}
\title{to check current SCOPE options}
\usage{
check_SCOPE_options(
  patch = "D:/model/STEMMUS_SCOPE/",
  site_name = NA,
  run_name = NA
)
}
\arguments{
\item{patch}{the patch to the STEMMUS_SCOPE model directory, default "D:/model/STEMMUS_SCOPE/",}

\item{site_name, run_name}{the name of the location and the name of run. The last can be used to name runs with different model parameters or settings,}
}
\value{
a table with information about SCOPE current selected model options
}
\description{
\code{check_SCOPE_options} present current SCOPE model options selected in the input_data.xlsx file
}
\details{
This function shows current SCOPE model options selected to run. If not changes were done
the options are the defaults @seealso \code{\link[=info_SCOPE_options]{info_SCOPE_options()}}
}
\examples{

\dontrun{
check_SCOPE_options(site_name = "DE-Hai", run_name = "ECdata_36")

formattable::formattable(check_SCOPE_options(site_name = "DE-Hai", run_name = "ECdata_36"),
                         align = c("c","l"))
}

}
\seealso{
Other check the current parameters and settings: 
\code{\link{change_SCOPE_constants}()},
\code{\link{change_SCOPE_options}()},
\code{\link{change_STEMMUS_DefineConstants}()},
\code{\link{change_STEMMUS_ModelSettings}()},
\code{\link{change_STEMMUS_SoilConstants}()},
\code{\link{check_SCOPE_constants}()},
\code{\link{check_SCOPE_tsInputs}()},
\code{\link{check_STEMMUS_DefineConstants}()},
\code{\link{check_STEMMUS_ForcingGlobal}()},
\code{\link{check_STEMMUS_ModelSettings}()},
\code{\link{check_STEMMUS_SoilConstants}()},
\code{\link{check_STEMMUS_SoilInitials}()},
\code{\link{check_STEMMUS_SoilProperties}()}
}
\concept{check the current parameters and settings}
