% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_.R
\name{check_SCOPE_tsInputs}
\alias{check_SCOPE_tsInputs}
\title{a function to check current SCOPE time series input files used by the model}
\usage{
check_SCOPE_tsInputs(
  patch = "D:/model/STEMMUS_SCOPE/",
  site_name = NA,
  run_name = NA
)
}
\arguments{
\item{patch}{the patch to the STEMMUS_SCOPE model directory, default "D:/model/STEMMUS_SCOPE/",}

\item{site_name, run_name}{the name of the location and the name of run. The last can be used to
name runs with different model parameters or settings,}
}
\value{
a table with information about SCOPE model current time series inputs
}
\description{
\code{check_SCOPE_tsInputs} present current time series inputs used by the SCOPE model
}
\details{
This function shows current time series inputs used by SCOPE to run the model. More information about the
variable desciption and units @seealso \code{\link[=info_SCOPE_tsInputs]{info_SCOPE_tsInputs()}}.
}
\examples{

check_SCOPE_ts_Inputs(site_name = "DE-Hai", run_name = "ECdata_36")

formattable::formattable(check_SCOPE_ts_Inputs(site_name = "DE-Hai", run_name = "ECdata_36"),
            align = c("l","l"))

}
\seealso{
Other check the current parameters and settings: 
\code{\link{change_SCOPE_constants}()},
\code{\link{change_SCOPE_options}()},
\code{\link{change_STEMMUS_DefineConstants}()},
\code{\link{change_STEMMUS_ModelSettings}()},
\code{\link{change_STEMMUS_SoilConstants}()},
\code{\link{check_SCOPE_constants}()},
\code{\link{check_SCOPE_options}()},
\code{\link{check_STEMMUS_DefineConstants}()},
\code{\link{check_STEMMUS_ForcingGlobal}()},
\code{\link{check_STEMMUS_ModelSettings}()},
\code{\link{check_STEMMUS_SoilConstants}()},
\code{\link{check_STEMMUS_SoilInitials}()},
\code{\link{check_STEMMUS_SoilProperties}()}
}
\concept{check the current parameters and settings}
