% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_.R
\name{check_STEMMUS_DefineConstants}
\alias{check_STEMMUS_DefineConstants}
\title{a function to check the current STEMMUS Model constants from the define file}
\usage{
check_STEMMUS_DefineConstants(patch = "D:/model/STEMMUS_SCOPE/")
}
\arguments{
\item{patch}{the patch to the STEMMUS_SCOPE model directory, default "D:/model/STEMMUS_SCOPE/",}
}
\value{
a table with information about STEMMUS current defined constants inputs
}
\description{
\code{check_STEMMUS_DefineConstants} present current model defined Constants inputs for STEMMUS
}
\details{
This function shows current STEMMUS model defined constants inputs. If not changes were done
the options are the defaults @seealso \code{\link[=info_STEMMUS_DefineConstants]{info_STEMMUS_DefineConstants()}}.
\strong{important} differently from SCOPE constant parameters any change in the settings will impact all
next runs and not a specific site_name/run_name run as in SCOPE. These constant are different from the
soil constant \code{\link[=info_STEMMUS_SoilConstants]{info_STEMMUS_SoilConstants()}}.
}
\examples{

formattable::formattable(check_STEMMUS_DefineConstants(), align = c("l","c"))

}
\seealso{
Other check the current parameters and settings: 
\code{\link{change_SCOPE_constants}()},
\code{\link{change_SCOPE_options}()},
\code{\link{change_STEMMUS_DefineConstants}()},
\code{\link{change_STEMMUS_ModelSettings}()},
\code{\link{change_STEMMUS_SoilConstants}()},
\code{\link{check_SCOPE_constants}()},
\code{\link{check_SCOPE_options}()},
\code{\link{check_SCOPE_tsInputs}()},
\code{\link{check_STEMMUS_ForcingGlobal}()},
\code{\link{check_STEMMUS_ModelSettings}()},
\code{\link{check_STEMMUS_SoilConstants}()},
\code{\link{check_STEMMUS_SoilInitials}()},
\code{\link{check_STEMMUS_SoilProperties}()}
}
\concept{check the current parameters and settings}
