% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_.R
\name{check_STEMMUS_ModelSettings}
\alias{check_STEMMUS_ModelSettings}
\title{a function to check the STEMMUS Model Settings options}
\usage{
check_STEMMUS_ModelSettings(patch = "D:/model/STEMMUS_SCOPE/")
}
\arguments{
\item{patch}{the patch to the STEMMUS_SCOPE model directory, default "D:/model/STEMMUS_SCOPE/",}
}
\value{
a table with information about STEMMUS current selected model settings or options
}
\description{
\code{check_STEMMUS_ModelSettings} present current model settings (options) for STEMMUS
}
\details{
This function shows current STEMMUS model options selected to run. If not changes were done
the options are the defaults @seealso \code{\link[=info_STEMMUS_ModelSettings]{info_STEMMUS_ModelSettings()}}.
\strong{important} differently from SCOPE options any change in the settings will impact all
next runs and not a specific site_name/run_name as in SCOPE.
}
\examples{

formattable::formattable(check_STEMMUS_ModelSettings(), align = c("l","c"))

}
\seealso{
Other check the current parameters and settings: 
\code{\link{change_SCOPE_constants}()},
\code{\link{change_SCOPE_options}()},
\code{\link{change_STEMMUS_DefineConstants}()},
\code{\link{change_STEMMUS_ModelSettings}()},
\code{\link{change_STEMMUS_SoilConstants}()},
\code{\link{check_SCOPE_constants}()},
\code{\link{check_SCOPE_options}()},
\code{\link{check_SCOPE_tsInputs}()},
\code{\link{check_STEMMUS_DefineConstants}()},
\code{\link{check_STEMMUS_ForcingGlobal}()},
\code{\link{check_STEMMUS_SoilConstants}()},
\code{\link{check_STEMMUS_SoilInitials}()},
\code{\link{check_STEMMUS_SoilProperties}()}
}
\concept{check the current parameters and settings}
