% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_.R
\name{check_STEMMUS_SoilConstants}
\alias{check_STEMMUS_SoilConstants}
\title{to check the current STEMMUS Model Soil Constants}
\usage{
check_STEMMUS_SoilConstants(patch = "D:/model/rSTEMMUS_SCOPE/")
}
\arguments{
\item{patch}{the patch to the STEMMUS_SCOPE model directory, default "D:/model/rSTEMMUS_SCOPE/",}
}
\value{
a table with information about STEMMUS current Soil Constants inputs
}
\description{
\code{check_STEMMUS_SoilConstants} present current model Soil Constants inputs for STEMMUS
}
\details{
This function shows current STEMMUS model Soil Constants inputs. If not changes were done
the options are the defaults @seealso \code{\link[=info_STEMMUS_SoilConstants]{info_STEMMUS_SoilConstants()}}.
\strong{important} differently from SCOPE constant parameters any change in the settings will impact all
next runs and not a specific site_name/run_name run as in SCOPE.
}
\examples{

\dontrun{
formattable::formattable(check_STEMMUS_SoilConstants(), align = c("l","c"))
}

}
\seealso{
Other check the current parameters and settings: 
\code{\link{change_SCOPE_constants}()},
\code{\link{change_SCOPE_options}()},
\code{\link{change_STEMMUS_DefineConstants}()},
\code{\link{change_STEMMUS_ModelSettings}()},
\code{\link{change_STEMMUS_SoilConstants}()},
\code{\link{check_SCOPE_constants}()},
\code{\link{check_SCOPE_options}()},
\code{\link{check_SCOPE_tsInputs}()},
\code{\link{check_STEMMUS_DefineConstants}()},
\code{\link{check_STEMMUS_ForcingGlobal}()},
\code{\link{check_STEMMUS_ModelSettings}()},
\code{\link{check_STEMMUS_SoilInitials}()},
\code{\link{check_STEMMUS_SoilProperties}()}
}
\concept{check the current parameters and settings}
