% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_.R
\name{check_STEMMUS_SoilInitials}
\alias{check_STEMMUS_SoilInitials}
\title{to check the current STEMMUS Model soil initial}
\usage{
check_STEMMUS_SoilInitials(
  patch = "D:/model/rSTEMMUS_SCOPE/",
  site_name = NA,
  run_name = NA
)
}
\arguments{
\item{patch}{the patch to the STEMMUS_SCOPE model directory, default "D:/model/rSTEMMUS_SCOPE/",}

\item{site_name, run_name}{the name of the location and the name of run. The last can be used to name runs with different model parameters or settings,}
}
\value{
a table with information about STEMMUS current soil initial inputs
}
\description{
\code{check_STEMMUS_SoilInitials} present current model soil initial inputs for STEMMUS
}
\details{
This function shows current STEMMUS model soil initial temperature and water content inputs in
different depths. These constants for the first timestamps are required to run the model @seealso \code{\link[=get_SoilInitials]{get_SoilInitials()}}.
\strong{important} the soil initial values are passed by the function @seealso \code{\link[=input_constants]{input_constants()}}
for each run.
}
\examples{

\dontrun{
formattable::formattable(check_STEMMUS_SoilInitials(site_name ="DE-Hai",
run_name = "ECdata_36")[[1]], align = c("l","c"))
formattable::formattable(check_STEMMUS_SoilInitials(site_name ="DE-Hai",
run_name = "ECdata_36")[[2]], align = c("l","c"))
}

}
\seealso{
Other check the current parameters and settings: 
\code{\link{change_SCOPE_constants}()},
\code{\link{change_SCOPE_options}()},
\code{\link{change_STEMMUS_DefineConstants}()},
\code{\link{change_STEMMUS_ModelSettings}()},
\code{\link{change_STEMMUS_SoilConstants}()},
\code{\link{check_SCOPE_constants}()},
\code{\link{check_SCOPE_options}()},
\code{\link{check_SCOPE_tsInputs}()},
\code{\link{check_STEMMUS_DefineConstants}()},
\code{\link{check_STEMMUS_ForcingGlobal}()},
\code{\link{check_STEMMUS_ModelSettings}()},
\code{\link{check_STEMMUS_SoilConstants}()},
\code{\link{check_STEMMUS_SoilProperties}()}
}
\concept{check the current parameters and settings}
