% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_.R
\name{check_STEMMUS_SoilProperties}
\alias{check_STEMMUS_SoilProperties}
\title{a function to check the current STEMMUS Model soil properties}
\usage{
check_STEMMUS_SoilProperties(
  patch = "D:/model/STEMMUS_SCOPE/",
  site_name = NA,
  run_name = NA
)
}
\arguments{
\item{patch}{the patch to the STEMMUS_SCOPE model directory, default "D:/model/STEMMUS_SCOPE/",}

\item{site_name, run_name}{the name of the location and the name of run. The last can be used to
name runs with different model parameters or settings,}
}
\value{
a table with information about STEMMUS current soil properties inputs
}
\description{
\code{check_STEMMUS_SoilProperties} present current model soil properties inputs for STEMMUS
}
\details{
This function shows current STEMMUS model soil properties inputs in different depths.
These constants are required to run the model @seealso \code{\link[=get_SoilProperties]{get_SoilProperties()}}.
\strong{important} the soil initial values are passed by the function @seealso \code{\link[=set_static_inputs]{set_static_inputs()}}
for each run.
}
\examples{

formattable::formattable(check_STEMMUS_SoilProperties(site_name ="DE-Hai", run_name = "ECdata_36"), align = c("l","c","c","c","c","c","c","l"))

}
\seealso{
Other check the current parameters and settings: 
\code{\link{change_SCOPE_constants}()},
\code{\link{change_SCOPE_options}()},
\code{\link{change_STEMMUS_DefineConstants}()},
\code{\link{change_STEMMUS_ModelSettings}()},
\code{\link{change_STEMMUS_SoilConstants}()},
\code{\link{check_SCOPE_constants}()},
\code{\link{check_SCOPE_options}()},
\code{\link{check_SCOPE_tsInputs}()},
\code{\link{check_STEMMUS_DefineConstants}()},
\code{\link{check_STEMMUS_ForcingGlobal}()},
\code{\link{check_STEMMUS_ModelSettings}()},
\code{\link{check_STEMMUS_SoilConstants}()},
\code{\link{check_STEMMUS_SoilInitials}()}
}
\concept{check the current parameters and settings}
