% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SoilProperty.R
\name{download_SoilProperty}
\alias{download_SoilProperty}
\title{to download a folder with maps of required soil properties to run the \code{get_SoilSroperties()}}
\usage{
download_SoilProperty(
  patch = "D:/model/rSTEMMUS_SCOPE/",
  doi = "10.5281/zenodo.15488066",
  timeout = 1000
)
}
\arguments{
\item{patch}{the path to the STEMMUS_SCOPE downloaded, example:\code{"D:/models/rSTEMMUS_SCOPE/"}.}

\item{doi}{the Zenodo doi of the SoilProperty folder (do not change)}

\item{timeout}{the file is 24 GB so you may need to increase it, default 1000}
}
\value{
It return a message: SoilProperty folder download at \code{"patch/input/"}.
}
\description{
\code{download_SoilProperty} download a folder nammed SoilProperty a the \verb{/input} diretory.
}
\details{
This is a function to download required model inputs (see ReadME 1.4 Soil Properties).
\strong{important} It is only needed once after install the package with \code{initial_setup()}
}
\examples{

\dontrun{
download_SoilProperty(patch = "D:/model/rSTEMMUS_SCOPE/")
}

}
