% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CanopyHeight.R
\name{get_CanopyHeight}
\alias{get_CanopyHeight}
\title{to download Global Canopy Height 10m raster cropped to the area of interest}
\usage{
get_CanopyHeight(site_name = NA, N_S_degree = NA, W_E_degree = NA, patch = NA)
}
\arguments{
\item{site_name}{name of the site,}

\item{N_S_degree}{latitude of the point of interest,}

\item{W_E_degree}{Longitude of the point of interest,}

\item{patch}{the patch to save/write the file}
}
\value{
an .tif raster image with 10m pixel for the area of interest in the chosen patch
}
\description{
\code{get_CanopyHeight} write an .tif raster image with 10m pixel for the area of interest.
}
\details{
This is a function to download from the Global Canopy Height (GCH) product a 10m raster,
and cropped to the area of interest.
}
\examples{

\dontrun{
# run to get Canopy height hc
library(curl)
loc_csv <- data.frame("Latitude" = round(sf::st_coordinates(locations$geometry)[,2],2),
                      "Longitude" = round(sf::st_coordinates(locations$geometry)[,1],2))

for all location
get_GCH(site_name = "Selke",
       N_S_degree = loc_csv$Latitude[1],
       W_E_degree = loc_csv$Longitude[1],
       patch = "D:/model/STEMMUS_SCOPE/input/input_for_input/Forcing_Globals/")

Global Canopy Height 2020
canopy_height <- terra::rast("D:/model/STEMMUS_SCOPE/input/input_for_input/GCH/GCH_tile_Selke.tif")
canopy_height <- terra::crop(canopy_height, selke_square_ext)

hc_points <- data.frame(terra::extract(canopy_height, locations))
names(hc_points) <- c("Name", "VH")
}

}
\seealso{
Other get input parameters: 
\code{\link{get_SoilInitials}()},
\code{\link{get_SoilProperties}()}
}
\concept{get input parameters}
