% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SoilInitials.R
\name{get_SoilInitials}
\alias{get_SoilInitials}
\title{Download soil initial condition from Copernicus ERA5 (Soil temperature and Volumetric Soil Water per depth)}
\usage{
get_SoilInitials(
  dataset = "reanalysis-era5-land",
  product = "reanalysis",
  format = "grib",
  area_box = NA,
  vars = NA,
  years = NA,
  months = NA,
  days = NA,
  times = NA,
  patch = NA,
  target_f = "download.grib",
  timeout_m = 50
)
}
\arguments{
\item{dataset}{= "reanalysis-era5-land"}

\item{product}{= "reanalysis"}

\item{format}{= "grib"}

\item{area_box}{= c(52.10, 9.75, 51.0, 11.65),}

\item{vars}{= c(}

\item{patch}{the patch to save/write the file}

\item{target_f}{= "download2.grib",}

\item{timeout_m}{= 50}

\item{year, month, day, times}{include the first day of the time series "YYYY", "MM", "DD" and "HH:mm"}
}
\value{
returns list with all soil properties required
}
\description{
\code{get_SoilInitials} returns Soil temperature and Volumetric Soil Water per depth required to run the model.
}
\details{
This is a function to Download the soil initial from Copernicus ERA5. Soil temperature and Volumetric Soil Water per depth
are required to run the model.
\strong{important} to run this function is needed to set a key to the keychain (ecmwfr) to have access to the
Copernicus platform.
}
\examples{

\dontrun{
library(ecmwfr)
set a key to the keychain
wf_set_key(key = "xxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxx")

get_SoilInitials(dataset = "reanalysis-era5-land",
                 product = "reanalysis",
                 format = "grib",
                 area_box = c(52.10, 9.75, 51.0, 11.65),
                 #site_name = NA,
                 vars = c("skin_temperature",
                          "soil_temperature_level_1",
                          "soil_temperature_level_2",
                          "soil_temperature_level_3",
                          "soil_temperature_level_4",
                          "volumetric_soil_water_layer_1",
                          "volumetric_soil_water_layer_2",
                          "volumetric_soil_water_layer_3",
                          "volumetric_soil_water_layer_4"),
                years = "2023",
                months = "06",
                days = "01",
                times = "00:00",
                path_file = "D:/model/STEMMUS_SCOPE/input/input_for_input/Initial_condition",
                target_f = "download2.grib",
                timeout_m = 50)

rast_init_Cond_01JUN23 <- terra::rast("D:/model/STEMMUS_SCOPE/input/input_for_input/Initial_condition/Init_Cond_01_JUN.grib")
}

}
\seealso{
Other get input parameters: 
\code{\link{get_CanopyHeight}()},
\code{\link{get_SoilProperties}()}
}
\concept{get input parameters}
