% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_.R
\name{info_IGBP_classes}
\alias{info_IGBP_classes}
\title{information about the valid bioma classes from IGBP land cover}
\usage{
info_IGBP_classes()
}
\value{
a table with the available landcover classes from IGBP
}
\description{
\code{info_IGBP_classes} present the valid landcover classes from IGBP
}
\details{
Only this LCLU classes are accepted. \strong{important} need to be written exactly like that.
The function has no arguments, just need to run info_IGBP_classes()
}
\examples{

formattable::formattable(info_IGBP_classes(), align = c("l","c","c","l"))

}
\seealso{
Other information about parameters and settings: 
\code{\link{info_SCOPE_README}()},
\code{\link{info_SCOPE_options}()},
\code{\link{info_SCOPE_tsInputs}()},
\code{\link{info_STEMMUS_DefineConstants}()},
\code{\link{info_STEMMUS_ModelSettings}()},
\code{\link{info_STEMMUS_SoilConstants}()}
}
\concept{information about parameters and settings}
