% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_.R
\name{info_SCOPE_options}
\alias{info_SCOPE_options}
\title{information about the SCOPE options}
\usage{
info_SCOPE_options()
}
\value{
a table with information about SCOPE options
}
\description{
\code{info_SCOPE_options} present the information about the SCOPE model options from the input_data.xlsx file
}
\details{
This is function shows all SCOPE model options, values and default settings.
The function has no arguments, just need to run info_SCOPE_options()
}
\examples{

\dontrun{
formattable::formattable(info_SCOPE_options(), align = c("l","c","c","l"))
 }

}
\seealso{
Other information about parameters and settings: 
\code{\link{info_IGBP_classes}()},
\code{\link{info_SCOPE_README}()},
\code{\link{info_SCOPE_tsInputs}()},
\code{\link{info_STEMMUS_DefineConstants}()},
\code{\link{info_STEMMUS_ModelSettings}()},
\code{\link{info_STEMMUS_SoilConstants}()}
}
\concept{information about parameters and settings}
