% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_.R
\name{info_SCOPE_tsInputs}
\alias{info_SCOPE_tsInputs}
\title{information about time series input files}
\usage{
info_SCOPE_tsInputs(
  patch = "D:/model/STEMMUS_SCOPE/",
  site_name = NA,
  run_name = NA
)
}
\arguments{
\item{patch}{the patch to the STEMMUS_SCOPE model directory, default "D:/model/STEMMUS_SCOPE/",}

\item{site_name, run_name}{the name of the location and the name of run. The last can be used to name runs with different model parameters or settings,}
}
\value{
a table with information about SCOPE time series inputs
}
\description{
\code{info_SCOPE_tsInputs} present information about the SCOPE timeseries data informed at input_data.xlsx file
}
\details{
This is function shows all time series data used in the SCOPE and which ones are
currently being used by the model
}
\examples{

\dontrun{
info_SCOPE_ts_Inputs(site_name = "DE-Hai", run_name = "ECdata_36")

formattable(info_SCOPE_ts_Inputs(site_name = "DE-Hai", run_name = "ECdata_36"), align = c("l","l"))
}

}
\seealso{
Other information about parameters and settings: 
\code{\link{info_IGBP_classes}()},
\code{\link{info_SCOPE_README}()},
\code{\link{info_SCOPE_options}()},
\code{\link{info_STEMMUS_DefineConstants}()},
\code{\link{info_STEMMUS_ModelSettings}()},
\code{\link{info_STEMMUS_SoilConstants}()}
}
\concept{information about parameters and settings}
