% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_.R
\name{info_STEMMUS_DefineConstants}
\alias{info_STEMMUS_DefineConstants}
\title{information about the STEMMUS define constants file}
\usage{
info_STEMMUS_DefineConstants()
}
\value{
a table with information about some STEMMUS constants
}
\description{
\code{info_STEMMUS_DefineConstants} present the information about the STEMMUS define constants
}
\details{
This is function shows some STEMMUS model constants and their default values.
The function has no arguments, just need to run info_STEMMUS_DefineConstants()
}
\examples{

formattable::formattable(info_STEMMUS_DefineConstants(), align = c("l","c","c","l"))

}
\seealso{
Other information about parameters and settings: 
\code{\link{info_IGBP_classes}()},
\code{\link{info_SCOPE_README}()},
\code{\link{info_SCOPE_options}()},
\code{\link{info_SCOPE_tsInputs}()},
\code{\link{info_STEMMUS_ModelSettings}()},
\code{\link{info_STEMMUS_SoilConstants}()}
}
\concept{information about parameters and settings}
