% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_.R
\name{info_STEMMUS_ModelSettings}
\alias{info_STEMMUS_ModelSettings}
\title{information about the STEMMUS Model Settings and Options}
\usage{
info_STEMMUS_ModelSettings()
}
\value{
a table with information about STEMMUS settings
}
\description{
\code{info_STEMMUS_ModelSettings} present the information about the STEMMUS model settings/options
}
\details{
This is function shows STEMMUS model settings/options, default values, units and description.
The function has no arguments, just need to run info_STEMMUS_ModelSettings()
}
\examples{

\dontrun{
formattable::formattable(info_ModelSettings(), align = c("l","c","c","l"))
}

}
\seealso{
Other information about parameters and settings: 
\code{\link{info_IGBP_classes}()},
\code{\link{info_SCOPE_README}()},
\code{\link{info_SCOPE_options}()},
\code{\link{info_SCOPE_tsInputs}()},
\code{\link{info_STEMMUS_DefineConstants}()},
\code{\link{info_STEMMUS_SoilConstants}()}
}
\concept{information about parameters and settings}
