% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_.R
\name{info_STEMMUS_SoilConstants}
\alias{info_STEMMUS_SoilConstants}
\title{information about the STEMMUS Soil Constants}
\usage{
info_STEMMUS_SoilConstants()
}
\value{
a table with information about STEMMUS Soil Constants
}
\description{
\code{info_STEMMUS_SoilConstants} present the information about the STEMMUS Soil Constants
}
\details{
This is function shows STEMMUS model Soil Constants, units, default values and description.
The function has no arguments, just need to run info_STEMMUS_SoilConstants()
}
\examples{

formattable::formattable(info_STEMMUS_SoilConstants(), align = c("l","c","c","l"))

}
\seealso{
Other information about parameters and settings: 
\code{\link{info_IGBP_classes}()},
\code{\link{info_SCOPE_README}()},
\code{\link{info_SCOPE_options}()},
\code{\link{info_SCOPE_tsInputs}()},
\code{\link{info_STEMMUS_DefineConstants}()},
\code{\link{info_STEMMUS_ModelSettings}()}
}
\concept{information about parameters and settings}
