% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_setup.R
\name{initial_setup}
\alias{initial_setup}
\title{to create a input folder with required files to run the model}
\usage{
initial_setup(patch = "D:/model/rSTEMMUS_SCOPE/")
}
\arguments{
\item{patch}{the path to the STEMMUS_SCOPE downloaded, example:\code{"D:/models/rSTEMMUS_SCOPE/"}.}
}
\value{
It return a message that shows the main path to run all the functions.
}
\description{
\code{initial_setup} create the file structure with the MATLAB codes and inputs in a directory of choice.
}
\details{
This is a function to prepare the STEMMUS_SCOPE (Matlab) codes and required  model inputs.
\strong{important} It is only needed once after install the package as initial setup
}
\examples{

\dontrun{
initial_setup(patch = "D:/model/rSTEMMUS_SCOPE/")
}

}
