% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_timeseries.R
\name{input_timeseries}
\alias{input_timeseries}
\title{to set all time series model inputs for a location}
\usage{
input_timeseries(
  patch = "D:/model/rSTEMMUS_SCOPE/",
  site_name = NA,
  run_name = NULL,
  data_frame = NULL,
  t_file = NULL,
  year_file = NULL,
  Ta_file = NULL,
  RH_file = NULL,
  ea_file = NULL,
  VPD_file = NULL,
  Rin_file = NULL,
  Rli_file = NULL,
  p_file = NULL,
  u_file = NULL,
  CO2_file = NULL,
  rain_file = NULL,
  z_file = NULL,
  tts_file = NULL,
  LAI_file = NULL,
  hc_file = NULL,
  Vcmax_file = NULL,
  Cab_file = NULL,
  LIDF_file = NULL
)
}
\arguments{
\item{patch}{the patch to the STEMMUS_SCOPE model directory, default "D:/model/rSTEMMUS_SCOPE/",}

\item{site_name, run_name}{the name of the location and the name of run. The last can be used to name runs with different model parameters or settings,}

\item{data_frame}{if all data input are into one data.frame with column names (t_, Ta_, RH_, u_, p_, rain_, Rin_, Rli_, LAI_, CO2_, ea_)}

\item{t_file}{required}

\item{year_file}{required}

\item{Ta_file}{required}

\item{RH_file}{required}

\item{ea_file}{required}

\item{VPD_file}{required}

\item{Rin_file}{required}

\item{Rli_file}{required}

\item{p_file}{required}

\item{u_file}{required}

\item{CO2_file}{optional}

\item{rain_file}{required}

\item{z_file}{optional}

\item{tts_file}{optional}

\item{LAI_file}{optional, but very important}

\item{hc_file}{optional}

\item{Vcmax_file}{optional}

\item{Cab_file}{optional}

\item{LIDF_file}{optional}
}
\value{
returns all files with static from STEMMUS and SCOPE will be write in the folder input/site_name_run_name_time
}
\description{
\code{input_timeseries} set all time series parameters required to run the model.
}
\details{
This is a function to prepare all time series input parameters to run the model.
\strong{important} @seealso \code{\link[=info_SCOPE_README]{info_SCOPE_README()}} to more information about which inputs are required or optional,
also the units must be used.
}
\examples{
\dontrun{
input_timeseries(patch = "D:/model/rSTEMMUS_SCOPE/",
              site_name = "DE-HoH",
              run_name = "ECdata_01",
              t_file =	ts_DWD$t_,
              year_file	= year(ts_DWD$timestamp),
              Rin_file	= ts_DWD$Rin_sun,
              Rli_file	= ts_DWD$Rli_,
              p_file	= ts_DWD$p_,
              Ta_file =	ts_DWD$Ta_,
              RH_file = ts_DWD$RH_,
              ea_file =	ts_DWD$ea_,
              u_file	= ts_DWD$u_,
              rain_file = ts_DWD$rain_/36000,
              VPD_file = ts_DWD$VPD_,
              tts_file	= tts_calc,             # optional
              CO2_file	= DE_ICOS$CO2_,         # optional
              LAI_file =	LAI500_Modis$LAI)
}

}
\seealso{
Other set input parameters: 
\code{\link{input_constants}()}
}
\concept{set input parameters}
