% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{to run the model in MATLAB (as time series simulation)}
\usage{
run(
  patch = "D:/model/rSTEMMUS_SCOPE/",
  site_name = NA,
  run_name = NA,
  cores = NULL
)
}
\arguments{
\item{patch}{the patch to the STEMMUS_SCOPE model directory, default "D:/model/rSTEMMUS_SCOPE/",}

\item{site_name, run_name}{the name of the location and the name of run. The last can be used to name runs with different model parameters or settings,}

\item{cores}{set 1,2,4,6,8 cores for ech run \strong{experimental}}
}
\value{
run the model in MATLAB that will return the resuls in the folder ~/output/site_name_run_name_time/
}
\description{
\code{run} opens the MATLAB, run the model and close it when is done
}
\details{
This is a function run the model and MATLAB using the input variables set before
\strong{important} it needs MATLAB installed in the machine and the functions
}
\examples{
\dontrun{

run_Matlab(patch = "D:/model/rSTEMMUS_SCOPE/",
              site_name = "DE-HoH",
              run_name = "ECdata_01")

n = 10
for (i in 1:10) {

run_Matlab(patch = "D:/model/STEMMUS_SCOPE/",
           site_name = site_names[i],
           run_name = "DWD_36")
           svMisc::progress(i, n, progress.bar = TRUE, init = T) # time delay
           Sys.sleep(1800) # time delay in seconds 7200/60
           if (i == n) message("Done!")
}

}

}
\seealso{
[setup_folder())],  @seealso [set_static_inputs())],  @seealso [set_ts_inputs()] be previously ran
with the same site_name and run_name

[setup_folder())]: R:setup_folder())
[set_static_inputs())]: R:set_static_inputs())
[set_ts_inputs()]: R:set_ts_inputs()
}
