% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_inMATLAB.R
\name{run_inMATLAB}
\alias{run_inMATLAB}
\title{to run the model in MATLAB (as time series simulation)}
\usage{
run_inMATLAB(
  patch = "D:/model/rSTEMMUS_SCOPE/",
  site_name = NA,
  run_name = NA,
  cores = NULL
)
}
\arguments{
\item{patch}{the patch to the STEMMUS_SCOPE model directory, default \code{"D:/model/rSTEMMUS_SCOPE/"},}

\item{site_name, run_name}{the name of the location and the name of run. The last can be used to name runs with different model parameters or settings,}

\item{cores}{set 1,2,4,6,8 cores for ech run \strong{experimental}}
}
\value{
run the model in MATLAB that will return the resuls in the folder \verb{~/output/site_name_run_name_time/}
}
\description{
\code{run_inMATLAB} opens the MATLAB, run the model and close it when is done
}
\details{
This is a function run the model and MATLAB using the input variables set before
\strong{important} it needs MATLAB installed in the machine and the functions
}
\examples{
\dontrun{

run_inMATLAB(patch = "D:/model/rSTEMMUS_SCOPE/",
              site_name = "DE-HoH",
              run_name = "ECdata_01")

n = 10
for (i in 1:10) {

run_inMATLAB(patch = "D:/model/rSTEMMUS_SCOPE/",
           site_name = site_names[i],
           run_name = "DWD_36")
           svMisc::progress(i, n, progress.bar = TRUE, init = T) # time delay
           Sys.sleep(1800) # time delay in seconds 7200/60
           if (i == n) message("Done!")
}

}

}
\seealso{
[new_run())],  @seealso [input_constants())],  @seealso [input_timeseries()] be previously ran
with the same site_name and run_name

[new_run())]: R:new_run())
[input_constants())]: R:input_constants())
[input_timeseries()]: R:input_timeseries()
}
