% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_static_inputs.R
\name{set_static_inputs}
\alias{set_static_inputs}
\title{Setting all statical input parameters and model setting required to run the model}
\usage{
set_static_inputs(
  patch = "D:/model/STEMMUS_SCOPE/",
  site_name = NA,
  run_name = NA,
  n_timestamps = NA,
  timestep_min = NA,
  elevation = NA,
  IGBP_veg_long = NA,
  initial_volumetric_soil_water = data.frame(swvl1 = NA, swvl2 = NA, swvl3 = NA, swvl4 =
    NA),
  initial_soil_temperature = data.frame(skt = NA, stl1 = NA, stl2 = NA, stl3 = NA, stl4 =
    NA),
  soil_property_list = NA,
  coef_n = data.frame(depth0 = NA, depth5 = NA, depth30 = NA, depth60 = NA, depth100 =
    NA, depth200 = NA),
  coef_alpha = data.frame(depth0 = NA, depth5 = NA, depth30 = NA, depth60 = NA, depth100
    = NA, depth200 = NA),
  saturatedk = data.frame(depth0 = NA, depth5 = NA, depth30 = NA, depth60 = NA, depth100
    = NA, depth200 = NA),
  ks0 = data.frame(depth0 = NA),
  residualMC = data.frame(depth0 = NA, depth5 = NA, depth30 = NA, depth60 = NA, depth100
    = NA, depth200 = NA),
  saturatedMC = data.frame(depth0 = NA, depth5 = NA, depth30 = NA, depth60 = NA, depth100
    = NA, depth200 = NA),
  porosity = data.frame(depth0 = NA, depth5 = NA, depth30 = NA, depth60 = NA, depth100 =
    NA, depth200 = NA),
  theta_s0 = data.frame(depth0 = NA),
  fieldMC = data.frame(depth0 = NA, depth5 = NA, depth30 = NA, depth60 = NA, depth100 =
    NA, depth200 = NA),
  FOClay = data.frame(depth4.5 = NA, depth16.6 = NA, depth49.29 = NA, depth82.9 = NA,
    depth138.3 = NA, depth229.6 = NA),
  MSOCarbon = data.frame(depth4.5 = NA, depth16.6 = NA, depth49.29 = NA, depth82.9 = NA,
    depth138.3 = NA, depth229.6 = NA),
  FOSand = data.frame(depth4.5 = NA, depth16.6 = NA, depth49.29 = NA, depth82.9 = NA,
    depth138.3 = NA, depth229.6 = NA),
  coef_Lamda = data.frame(l1 = NA, l3 = NA, l5 = NA, l6 = NA, l7 = NA, l8 = NA),
  fmax = NA,
  Simulation_Name = site_name,
  soil_file = "soilnew.txt",
  leaf_file = "Optipar2017_ProspectD.mat",
  atmos_file = "FLEX-S3_std.atm",
  setoptions = c(1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1),
  veg_type_C = 3,
  Cab = 120,
  Cca = 30,
  Cdm = 0.015,
  Cw = 0.009,
  Cs = 0,
  Cant = 0,
  N = 1.4,
  rho_thermal = 0.01,
  tau_thermal = 0.01,
  Vcmo = 80,
  m = 9,
  BallBerry0 = 0.015,
  Type = 0,
  kV = 0.6396,
  Rdparam = 0.015,
  Tparam = c(0.2, 0.3, 288, 313, 328),
  Tyear = 9.74,
  beta = 0.507,
  kNPQs = 0,
  qLs = 1,
  stressfactor = 1,
  fqe = 0.01,
  spectrum = 1,
  rss = 500,
  rs_thermal = 0.06,
  cs = 1180,
  rhos = 1800,
  lambdas = 1.55,
  SMC = 25,
  BSMBrightness = 0.5,
  BSMlat = LAT,
  BSMlon = LON,
  LAI = 3,
  hc = 25,
  LIDFa = -0.35,
  LIDFb = -0.15,
  leafwidth = 0.1,
  z = 40,
  Rin = 600,
  Ta = 10,
  Rli = 300,
  p = 970,
  ea = 15,
  u = 2,
  Ca = 410,
  Oa = 209,
  zo = 2.5,
  d = 5.34,
  Cd = 0.3,
  rb = 10,
  CR = 0.35,
  CD1 = 20.6,
  Psicor = 0.2,
  CSSOIL = 0.01,
  rbs = 10,
  rwc = 0,
  startDOY = 0,
  endDOY = 366,
  LAT = LAT,
  LON = LON,
  timezn = 1,
  tts = 30,
  tto = 0,
  psi = 90,
  delHaV = 65330,
  delSV = 485,
  delHdV = 149250,
  delHaJ = 43540,
  delSJ = 495,
  delHdJ = 152040,
  delHaP = 53100,
  delSP = 490,
  delHdP = 150650,
  delHaR = 46390,
  delSR = 490,
  delHdR = 150650,
  delHaKc = 79430,
  delHaKo = 36380,
  delHaT = 37830,
  Q10 = 2,
  s1 = 0.3,
  s2 = 313.15,
  s3 = 0.2,
  s4 = 288.15,
  s5 = 1.3,
  s6 = 328.15
)
}
\arguments{
\item{patch}{the patch to the STEMMUS_SCOPE model directory, default "D:/model/STEMMUS_SCOPE/",}

\item{site_name, run_name}{the name of the location and the name of run. The last can be used to name runs with different model parameters or settings,}

\item{timestep_min, }{four hourly data use 60, half hour 30 (maximum possible is 180 for timestamp of every 3 hours)}

\item{IGBP_veg_long}{the main IGBP landcover class in the point of interest,}

\item{initial_volumetric_soil_water}{volumetric soil water in the 6 depths (). It can be obtained in ERA5 data with funtion @seealso \code{\link[=get_SoilInitials]{get_SoilInitials()}}}

\item{initial_soil_temperature, }{soil temperature in the 6 depths (). It can be obtained in ERA5 data with funtion get_SoilInitials}

\item{soil_property_list}{a list of soil properties extracted from the function @seealso \code{\link[=get_SoilPropertie]{get_SoilPropertie()}}}

\item{coef_n, coef_alpha, saturatedk, ks0, residualMC, saturatedMC, porosity, theta_s0, fieldMC, FOClay, MSOCarbon, FOSand, coef_Lamda, fmax}{alternatively
include a data.frame("depth0"=X,"depth5"=X,"depth30"=X,"depth60"=X,"depth100"=X,"depth200"=X) with the values for each soil property,
the inputs ks0, theta_s0 and fmax require only one value. @seealso \code{\link[=get_soil_properties]{get_soil_properties()}}}

\item{soil_file, leaf_file, atmos_file, setoptions}{default "soilnew.txt", "Optipar2017_ProspectD.mat","FLEX-S3_std.atm",  @seealso \code{\link[=info_SCOPE_README(), info_SCOPE_options]{info_SCOPE_README(), info_SCOPE_options()}}
SCOPE model inputs, constants from inputdata spreadsheet}

\item{setoptions}{a vector of model settings from the input_data.xls options spreadsheet @seealso \code{\link[=info_SCOPE_options]{info_SCOPE_options()}}}

\item{veg_type_C}{default is 3, alternately can be use 4}

\item{Cab, Cca, Cdm, Cw, Cs, Cant, N, rho_thermal, tau_thermal}{PROSPECT submodel parameters for the default values @seealso \code{\link[=get_SCOPE_constants]{get_SCOPE_constants()}}}

\item{Vcmo, m, BallBerry0, Type, kV, Rdparam, Tparam, Tyear, beta, kNPQs, qLs, stressfactor}{Leaf_Biochemical parameters for the default values @seealso \code{\link[=get_SCOPE_constants]{get_SCOPE_constants()}}}

\item{fqe}{Fluorescence parameters @seealso \code{\link[=get_SCOPE_constants]{get_SCOPE_constants()}}}

\item{spectrum, rss, rs_thermal, cs, rhos, lambdas, SMC, BSMBrightness, BSMlat, BSMlon}{Soil parameters @seealso \code{\link[=get_SCOPE_constants]{get_SCOPE_constants()}}}

\item{LAI, hc, LIDFa, LIDFb, leafwidth}{Canopy parameters @seealso \code{\link[=get_SCOPE_constants]{get_SCOPE_constants()}}}

\item{hc}{canopy height in the point of interest,}

\item{z, Rin, Rli, Ta, p, ea, u, Ca, Oa}{Meteo parameters @seealso \code{\link[=get_SCOPE_constants]{get_SCOPE_constants()}}}

\item{zo, d, Cd, rb, CR, CD1, Psicor, CSSOIL, rbs, rwc}{Aerodynamic parameters @seealso \code{\link[=get_SCOPE_constants]{get_SCOPE_constants()}}}

\item{startDOY, endDOY, timezn, n_timestamps}{day of the year for the start and end of the run period. The the timezone and number of timestamps being run.}

\item{tts, tto, psi}{Angles parameters @seealso \code{\link[=get_SCOPE_constants]{get_SCOPE_constants()}}}

\item{delHaV, delSV, delHdV, delHaJ, delSJ, delHdJ, delHaP, delSP, delHdP, delHaR, delSR, delHdR, delHaKc, delHaKo, delHaT, Q10, s1, s2, s3, s4, s5, s6}{Photosynthetic Temperature Dependence Functional parameters @seealso \code{\link[=get_SCOPE_constants]{get_SCOPE_constants()}}}

\item{lat, lon, elevation}{the latitude, longitude and elevation of the point of interest,}
}
\value{
returns all files with static from STEMMUS and SCOPE will be write in the folder input/site_name_run_name_time
}
\description{
\code{set_static_inputs} set all statical parameters and model setting required to run the model.
}
\details{
This is a function to prepare all static input parameters to run the model.
\strong{important} the result of the get_SoilProperty needs to be informed,
otherwise all the soil properties need to be included manually.
}
\examples{
\dontrun{
set_static_inputs(patch = "D:/model/STEMMUS_SCOPE/",
                  site_name = "mySite",
                  run_name = "grass_2019",
                  LAT = 37.83,
                  LON = 107.69,
                  elevation = 500,
                  IGBP_veg_long = "Grasslands",
                  hc = 0.03,
                  z = 2,
                  zo = 0.00369,
                  d = 0.0201,
                  n_timestamps = 7344, # number of timestamps
                  timestep_min = 30, # hourly 60, half hour 3
                  initial_soil_temperature = data.frame("skt"=Initial_01MAY19_Yachi_p[1,2],
                                                        "stl1"=Initial_01MAY19_Yachi_p[2,2],
                                                        "stl2"=Initial_01MAY19_Yachi_p[3,2],
                                                        "stl4"=Initial_01MAY19_Yachi_p[5,2]), #CDS ERA5Land layers 1 to 4
                  initial_volumetric_soil_water = data.frame("swvl1"=Initial_01MAY19_Yachi_p[6,2],
                                                             "swvl2"=Initial_01MAY19_Yachi_p[7,2],
                                                             "swvl3"=Initial_01MAY19_Yachi_p[8,2],
                                                             "swvl4"=Initial_01MAY19_Yachi_p[9,2]), #CDS ERA5Land layers 1 to 4
                  soil_property_list = Soil_property_CRNs,
                  startDOY = 121,
                  endDOY = 274,
                  timezn = 8,
                  Tyear = 8.3,
                  Cab = 80,
                  Cca = 25,
                  Cdm = 0.012,
                  LAI = 0.46,
                  Vcmo = 120,
                  m = 7,
                  BallBerry0 = 0.025,
                  Ta = 20,
                  SMC = 0.15,
                  LIDFa = -0.4667,
                  leafwidth = 0.003,
                  Ca = 380,
                  setoptions = c(1,1,1,0,0,1,0,0,1,0,1,0,1,1,0,1,0,1))
}

}
\seealso{
Other set input parameters: 
\code{\link{set_ts_inputs}()}
}
\concept{set input parameters}
